Privacy_Handler_for_LLMs/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
faker==19.3.0
python-dotenv==1.0.0
numpy==1.26.4
gunicorn==21.2.0
google-generativeai==0.3.2